+++ packages/lumx-react/src/components/text-field/TextField.stories.tsx
@@ -57,3 +57,55 @@ export const textFieldWithError = ({ theme }: any) => (
         }
     />
 );
+
+export const TextArea = ({ theme }: any) => {
+    const [value, setValue] = React.useState('my value');
+    return (
+        <TextField
+            value={value}
+            label={text('Label', 'I am the label')}
+            placeholder={text('Placeholder', 'ex: A value')}
+            multiline
+            minimumRows={1}
+            theme={theme}
+            onChange={setValue}
+            helper={<span>{text('Helper', 'ex: toto@acme.com')}</span>}
+        />
+    );
+};
+export const TextAreaWith2Lines = ({ theme }: any) => {
+    const [value, setValue] = React.useState('the value');
+    return (
+        <TextField
+            value={value}
+            label={text('Label', 'I am the label')}
+            placeholder={text('Placeholder', 'ex: A value')}
+            multiline
+            minimumRows={2}
+            theme={theme}
+            onChange={setValue}
+            helper={<span>{text('Helper', 'ex: toto@acme.com')}</span>}
+        />
+    );
+};
+
+export const TextAreaWithManyLines = ({ theme }: any) => {
+    const myvalue = `I
+am
+a
+multiline
+text`;
+    const [value, setValue] = React.useState(myvalue);
+    return (
+        <TextField
+            value={value}
+            label={text('Label', 'I am the label')}
+            placeholder={text('Placeholder', 'ex: A value')}
+            multiline
+            minimumRows={2}
+            theme={theme}
+            onChange={setValue}
+            helper={<span>{text('Helper', 'ex: toto@acme.com')}</span>}
+        />
+    );
+};