+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `TextField` component is now computing the number of rows (if `multiple` is set) at the first rendering
+
 ## [0.23.0][] - 2020-05-29
 
 ### Added